pyproject.toml
[tool.poetry]
name = "search"
version = "0.1.0"
description = ""
authors = ["Jinsoo Kim <jjambbongjoa@gmail.com>"]

# ✅ 패키지 경로 명시 (Poetry가 패키지를 찾도록 설정)
packages = [{include = "app"}]

[tool.poetry.group.dev.dependencies]
pydevd-pycharm = "243.25659.43"

[tool.poetry.scripts]
scheduler = "scheduler:main"
api = "api:main"
command = "command:main"

[tool.poetry.dependencies]
python = "^3.12.9"
uvicorn = "^0.19.0"
pymongo = "4.12.0"
pydantic = {extras = ["email"], version = "^1.10.9"}
python-dotenv = "^0.21.0"
cryptography = "^38.0.3"
validators = "^0.20.0"
python-dateutil = "^2.8.2"
httpx = "^0.23.0"
pybase64 = "^1.2.3"
bcrypt = "^4.0.1"
import-ipynb = "^0.1.4"
dependency-injector = "^4.41.0"
notebook = "^6.5.4"
jupyter = "^1.0.0"
sqlalchemy = "^2.0.15"
pymysql = "^1.0.3"
pytz = "^2023.3"
click = "^8.1.3"
pydash = "^7.0.4"
numpy = "^1.26.4"
pandas = "^2.2.1"
pytest = "^8.1.1"
platformdirs = "^4.2.0"
fastjsonschema = "^2.19.1"
fastapi = "0.115.0"
pyarrow = "^16.1.0"
fastparquet = "^2024.5.0"
openpyxl = "^3.1.3"
apscheduler = "^3.10.4"
boto3 = "^1.34.144"
boto3-stubs = {extras = ["logs"], version = "^1.34.145"}
boto3-stubs-lite = {extras = ["essential"], version = "^1.34.145"}
gspread = "^6.1.2"
oauth2client = "^4.1.3"
google-analytics-data = "^0.18.11"
gspread-dataframe = "^4.0.0"
plotly = "^5.24.0"
ace-tools = "^0.0"
shapely = "^2.0.6"
geopandas = "^1.0.1"
geojson = "^3.1.0"
opensearch-py = "^2.8.0"
aws-requests-auth = "^0.4.3"
requests-aws4auth = "^1.3.1"
str2bool = "^1.1"
mysql-connector-python = "^9.4.0"
rich = "^14.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"